refactor: 공통 컴포넌트를 components/common로 이동
