+++ src/components/common/BaseImageInput.jsx
